Rename files PathPaymentOpFrame to PathPaymentStrictReceiveOpFrame
