+++ src/transactions/PathPaymentStrictReceiveOpFrame.h
